finish refacoring FE proejct view
